refactor: PieceParser의 네이밍 개선
